Добавил операции по манипулиции с сущностями * CRUD на запчасти * Добавил DTO'шки для запросов + для результатов * Конвертеры для DTO -> Entity && Entity -> DTO * Добавил  ExceptionHandler * Немного переработал контроллеры

+++ src/main/java/com/warehouse/ladaparts/dto/model/PartCartDTO.java
@@ -1,17 +1,18 @@
 package com.warehouse.ladaparts.dto.model;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 public class PartCartDTO extends PartDTO {
-    private Integer weight;
+    private BigDecimal weight;
     private List<ModelDTO> models;
     private List<PartDTO> compositionParts;
 
-    public Integer getWeight() {
+    public BigDecimal getWeight() {
         return weight;
     }
 
-    public void setWeight(Integer weight) {
+    public void setWeight(BigDecimal weight) {
         this.weight = weight;
     }
